% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\name{atributoDataset}
\alias{atributoDataset}
\title{Basic constructor of an AtributoDataset Object}
\usage{
atributoDataset(name, vector, cortes = NULL)
}
\arguments{
\item{name}{A character string representing the name of the attribute.}

\item{vector}{A vector of values for the attribute. The type of this vector determines the 'tipo' of the attribute.}

\item{cortes}{An optional parameter that can contain information about cuts or discretization applied to the values. Default is NULL.}
}
\value{
An object of class \code{\linkS4class{AtributoDataset}} containing the name, type, values, and cuts of the attribute.
}
\description{
This function creates an instance of the s \code{\linkS4class{AtributoDataset}} class.
}
