% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\name{calculateMetrics,Dataset-method}
\alias{calculateMetrics,Dataset-method}
\title{Calculate Metrics for Dataset Objects}
\usage{
\S4method{calculateMetrics}{Dataset}(obj)
}
\arguments{
\item{obj}{An object of class \code{\linkS4class{Dataset}} containing
a list of attributes for which the metrics have to be calculated.}
}
\value{
A list of named vectors, where each vector contains the calculated
metrics (Variance, AUC, and Entropy) for each attribute in the
\code{\linkS4class{Dataset}} object.
}
\description{
This method calculates various metrics for each attribute in an object of
class \code{\linkS4class{Dataset}}. The metrics calculated include variance,
area under the curve (AUC), and entropy, depending on the type of each
attribute.
}
\details{
- For numeric and integer attributes, the variance and AUC are calculated.
- For factor and logical attributes, entropy is calculated.

Note: The AUC is calculated based on the trapezoidal rule.
}
\examples{
numeric_attr <- atributoDataset("Numeric Attribute", c(1, 2, 3, 4, 5))
logical_attr <- atributoDataset("Logical Attribute", c(TRUE, FALSE, TRUE, FALSE, TRUE))
dataset <- createDataset("My Dataset", list(numeric_attr, logical_attr))
calculateMetrics(dataset)

}
