% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\name{calculateRates,Dataset-method}
\alias{calculateRates,Dataset-method}
\title{Calculate True Positive Rate (TPR) and False Positive Rate (FPR)}
\usage{
\S4method{calculateRates}{Dataset}(obj, numeric_attr_name)
}
\arguments{
\item{obj}{An object of class \code{\linkS4class{Dataset}}.}

\item{numeric_attr_name}{A string which indicates the name of the numeric attribute
to be used for ordering the dataset.}
}
\value{
A matrix where each row contains the TPR and FPR for different cut points.
}
\description{
This method calculates the TPR and FPR for each cut point in AttributeDataset values.
}
\examples{
numeric_attr <- atributoDataset("Numeric Attribute", c(1, 2, 3, 4, 5))
logical_attr <- atributoDataset("Logical Attribute", c(TRUE, FALSE, TRUE, FALSE, TRUE))
my_dataset <- createDataset("My Dataset", list(numeric_attr, logical_attr))
calculateRates(my_dataset, "Numeric Attribute")
}
