% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\name{correlation,Dataset-method}
\alias{correlation,Dataset-method}
\title{Calculate Correlation and Mutual Information for a \code{\linkS4class{Dataset}}}
\usage{
\S4method{correlation}{Dataset}(obj)
}
\arguments{
\item{obj}{A \code{\linkS4class{Dataset}} object containing attributes to be analyzed.}
}
\value{
A list containing two matrices:
\itemize{
  \item \code{correlations}: A symmetric matrix of correlations for numeric attributes
  \item \code{mutual_information}: A symmetric matrix of mutual information values for categorical attributes
}
}
\description{
This method iterates trough all the atributes of a dataset and calculates the correlation (or mutual Information) for each pair.
For those calculations uses internal functions
}
\details{
This method calculates the correlation for numeric attributes and mutual information for categorical attributes
within a \code{\linkS4class{Dataset}} object. It returns two matrices: one for correlations and another for mutual information.
}
\examples{
dataset <- createDataset(nombre = "Example", atributos = list(
  atributoDataset("Numeric1", c(1, 2, 3)),
  atributoDataset("Numeric2", c(4, 4, 6))))
result <- correlation(dataset)
correlation_matrix <- result$correlations
mutual_information_matrix <- result$mutual_information

}
