% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\name{createDataset}
\alias{createDataset}
\title{Create a Dataset Object}
\usage{
createDataset(nombre, atributos)
}
\arguments{
\item{nombre}{A character string representing the name of the dataset.}

\item{atributos}{A list of objects of class AtributoDataset representing the attributes of the dataset.}
}
\value{
An object of class Dataset containing the name and attributes.
}
\description{
This function creates an instance of the Dataset class.
}
