% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\docType{class}
\name{Dataset-class}
\alias{Dataset-class}
\title{Dataset Class}
\description{
This class represents a dataset containing a name and a list of attributes.
}
\section{Slots}{

\describe{
\item{\code{nombre}}{A character string representing the name of the dataset.}

\item{\code{atributos}}{A list of objects of class AtributoDataset, representing the attributes of the dataset.}
}}

