% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\name{datasetFromCSV}
\alias{datasetFromCSV}
\title{Create a Dataset Object from a CSV File}
\usage{
datasetFromCSV(file_path, sep = ",")
}
\arguments{
\item{file_path}{A character string representing the path to the CSV file.}

\item{sep}{A character string specifying the separator used in the CSV file (default is a comma).}
}
\value{
An object of class \code{\linkS4class{Dataset}} containing the name derived from the CSV filename and the attributes represented as \code{\linkS4class{AtributoDataset}} objects.
}
\description{
This function reads a CSV file and creates an instance of the \code{\linkS4class{Dataset}} class.
}
