% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\name{discretizeEF,AtributoDataset-method}
\alias{discretizeEF,AtributoDataset-method}
\title{Discretize AtributoDataset Using Equal Frequency}
\usage{
\S4method{discretizeEF}{AtributoDataset}(obj, n_intervals)
}
\arguments{
\item{obj}{An object of class \code{\linkS4class{AtributoDataset}} to be discretized.}

\item{n_intervals}{An integer specifying the number of intervals to create for discretization.}
}
\value{
A new \code{\linkS4class{AtributoDataset}} object containing the discretized values and cut points.
}
\description{
This method discretizes the numeric or integer values of an \code{\linkS4class{AtributoDataset}} object into specified intervals using equal frequency
Thanks to this discretizing technique all intervals have the same amount of items +-1
In the case that the vector contains duplicated values, sometimes it will be imposible to divide all intervals perfectly.
So, if the cut point is a duplicated value in the vector, we have decided to add those duplicated values into the lower interval
}
\examples{
# Create an AtributoDataset object
vector <- c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10)
atributo <- atributoDataset("Example Attribute", vector)

# Discretize the AtributoDataset object into 3 intervals
discretized_atributo <- discretizeEF(atributo, n_intervals = 3)

}
