% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\name{discretizeEW,AtributoDataset-method}
\alias{discretizeEW,AtributoDataset-method}
\title{Discretize AtributoDataset Using Equal Width}
\usage{
\S4method{discretizeEW}{AtributoDataset}(obj, n_intervals)
}
\arguments{
\item{obj}{An object of class \code{\linkS4class{AtributoDataset}} to be discretized.}

\item{n_intervals}{An integer specifying the number of intervals to create for discretization.}
}
\value{
A new \code{AtributoDataset} object containing the discretized values and cut points.
}
\description{
This method discretizes the numeric or integer values of an \code{\linkS4class{AtributoDataset}} object into specified intervals using equal width.
Equal width means all intervals have the same length
}
\examples{
# Create an AtributoDataset object
vector <- c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10)
atributo <- atributoDataset("Test", vector)

# Discretize the AtributoDataset object into 3 intervals
discretized_atributo <- discretizeEW(atributo, 3)

# View the discretized values and cut points
print(discretized_atributo)

}
