% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\name{entropy,AtributoDataset-method}
\alias{entropy,AtributoDataset-method}
\title{Calculate Entropy for AtributoDataset Objects}
\usage{
\S4method{entropy}{AtributoDataset}(obj)
}
\arguments{
\item{obj}{An object of class \code{\linkS4class{AtributoDataset}}
containing the values (logical or factor) for which the entropy is to be calculated.
If values are numeric, it will throw an error}
}
\value{
A numeric value representing the entropy of the values in the
specified \code{\linkS4class{AtributoDataset}} object.
}
\description{
This method calculates the entropy of the values contained in an
object of class \code{\linkS4class{AtributoDataset}}. Entropy is a measure
of the uncertainty or randomness in the data.
}
\examples{
# Create an instance of AtributoDataset
atributo <- atributoDataset("example", factor(c("blue","green","blue")))

entropy(atributo)
}
