% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\name{filterMetrics,Dataset-method}
\alias{filterMetrics,Dataset-method}
\title{Filter the attributes of a \code{\linkS4class{Dataset}} object based on specific metrics.}
\usage{
\S4method{filterMetrics}{Dataset}(obj, metric_name, operator, value)
}
\arguments{
\item{obj}{An object of class \code{\linkS4class{Dataset}} containing the attributes to be filtered.}

\item{metric_name}{A \code{character} string specifying the name of the metric to be used for filtering.
Only 'Variance' 'AUC' and 'Entropy' are allowed}

\item{operator}{A \code{character} string representing the comparison operator to be used.
Must be one of: \code{">"}, \code{"<"}, \code{">="}, \code{"<="}, \code{"=="}, \code{"!="}.}

\item{value}{A numeric value to compare the attribute metrics against.}
}
\value{
An object of class \code{\linkS4class{Dataset}} containing only the attributes that passed the filter.
}
\description{
This method allows filtering the attributes of a \code{Dataset} object according to the value
of a specified metric using a given operator. Attributes that do not meet the specified
condition will be removed from the object.
}
\examples{
dataset <- createDataset(nombre = "Example", atributos = list(
  atributoDataset("Months", c(1, 2, 3)),
  atributoDataset("FactorAttr", factor(c("A", "B", "A")))))
filtered_dataset <- filterMetrics(dataset, metric_name = "AUC", operator = ">", value = 0.5)
}
