% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\name{normalize,AtributoDataset-method}
\alias{normalize,AtributoDataset-method}
\title{Normalize Values for AtributoDataset Objects}
\usage{
\S4method{normalize}{AtributoDataset}(obj)
}
\arguments{
\item{obj}{An object of class \code{\linkS4class{AtributoDataset}}
containing the values to be normalized.}
}
\value{
An object of class \code{\linkS4class{AtributoDataset}} with the
normalized values and a modified name indicating normalization.
}
\description{
This method normalizes the values contained in an object of class
\code{\linkS4class{AtributoDataset}}. Normalization is performed only
for numeric or integer attributes.
}
\details{
If the attribute type is not numeric or integer, an error is thrown.
}
\examples{
atributo <- atributoDataset(name = "Attribute1", vector = c(1, 2, 3, 4, 5))
normalize(atributo)
}
