% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\name{normalize,Dataset-method}
\alias{normalize,Dataset-method}
\title{Normalize Attributes for Dataset Objects}
\usage{
\S4method{normalize}{Dataset}(obj)
}
\arguments{
\item{obj}{An object of class \code{\linkS4class{Dataset}} containing a list
of attributes to be normalized.}
}
\value{
An object of class \code{\linkS4class{Dataset}} with the normalized
numeric and integer attributes, while other attributes remain the same
}
\description{
This method normalizes the numeric and integer attributes contained in an
object of class \code{\linkS4class{Dataset}}
}
\examples{
dataset <- createDataset(nombre = "Example", atributos = list(
  atributoDataset(name = "Years", vector = c(1, 2, 3)),
  atributoDataset(name = "FactorAttr", vector = factor(c("A", "B", "A")))))
normalize(dataset)
}
