% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\name{standarize,AtributoDataset-method}
\alias{standarize,AtributoDataset-method}
\title{Standardize Values for AtributoDataset Objects}
\usage{
\S4method{standarize}{AtributoDataset}(obj)
}
\arguments{
\item{obj}{An object of class \code{\linkS4class{AtributoDataset}}
containing values to standardize.}
}
\value{
An object of class \code{\linkS4class{AtributoDataset}} with
standardized values.
}
\description{
This method standardizes the values in an object of class
\code{\linkS4class{AtributoDataset}}. It can only be applied to numeric
or integer attributes.
}
\details{
If the attribute type is not numeric or integer, an error is thrown
}
\examples{
atributo <- atributoDataset(name = "Attribute1", vector = c(1, 2, 3, 4, 5))
standarize(atributo)
}
