% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataset.R
\name{variance,AtributoDataset-method}
\alias{variance,AtributoDataset-method}
\title{Calculate Variance of Numeric Attributes}
\usage{
\S4method{variance}{AtributoDataset}(obj)
}
\arguments{
\item{obj}{An object of class \code{\linkS4class{AtributoDataset}}.}
}
\value{
A numeric value representing the variance of the attribute's values.
}
\description{
This method calculates the variance of the numeric values stored in an
object of class \code{\linkS4class{AtributoDataset}}. If the attribute
is not numeric or integer, it will throw an error
}
\examples{
# Creating a numeric attribute
numeric_attr <- atributoDataset("Numeric Attribute", c(1, 2, 3, 4, 5))
variance(numeric_attr)

}
